added `findById` method on both DAO and service layers

+++ src/main/java/com/techdegree/instateam/dao/CollaboratorDaoImpl.java
@@ -24,7 +24,7 @@ public class CollaboratorDaoImpl implements CollaboratorDao {
     }
 
     @Override
-    public Collaborator findById(Long id) {
+    public Collaborator findById(int id) {
         return null;
     }
